final alignment w/ canonical example; disable progress bar and logging
