+++ examples/quickstart-fastai/fastai_example/server_app.py
@@ -2,9 +2,12 @@
 
 from typing import List, Tuple
 
-from flwr.common import Context, Metrics
+from fastai.vision.all import squeezenet1_1
+from flwr.common import Context, Metrics, ndarrays_to_parameters
 from flwr.server import ServerApp, ServerAppComponents, ServerConfig
 from flwr.server.strategy import FedAvg
+
+from fastai_example.task import get_params
 
 
 # Define metric aggregation function
@@ -20,12 +23,23 @@ def weighted_average(metrics: List[Tuple[int, Metrics]]) -> Metrics:
 
 def server_fn(context: Context) -> ServerAppComponents:
     """Construct components for ServerApp."""
-    # Construct ServerConfig
-    num_rounds = context.run_config["num_server_rounds"]
-    config = ServerConfig(num_rounds=num_rounds)
+
+    # Let's define the global model and pass it to the strategy
+    # Note this is optional.
+    parameters = ndarrays_to_parameters(get_params(squeezenet1_1()))
 
     # Define strategy
-    strategy = FedAvg(evaluate_metrics_aggregation_fn=weighted_average)
+    fraction_fit = context.run_config["fraction-fit"]
+    strategy = FedAvg(
+        fraction_fit=fraction_fit,
+        fraction_evaluate=0.5,
+        evaluate_metrics_aggregation_fn=weighted_average,
+        initial_parameters=parameters,
+    )
+
+    # Construct ServerConfig
+    num_rounds = context.run_config["num-server-rounds"]
+    config = ServerConfig(num_rounds=num_rounds)
 
     return ServerAppComponents(config=config, strategy=strategy)
