+++ examples/quickstart-fastai/pyproject.toml
@@ -3,13 +3,13 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
-name = "quickstart-fastai"
-version = "0.1.0"
-description = "Fastai Federated Learning Quickstart with Flower"
-authors = [{ name = "The Flower Authors", email = "hello@flower.ai" }]
+name = "fastai_example"
+version = "1.0.0"
+description = "Federated Learning with Fastai and Flower (Quickstart Example)"
+license = "Apache-2.0"
 dependencies = [
-    "flwr[simulation]>=1.10.0,<2.0",
-    "flwr-datasets[vision]>=0.0.3,<1.0.0",
+    "flwr[simulation]>=1.10.0",
+    "flwr-datasets[vision]>=0.3.0",
     "fastai==2.7.14",
     "torch==2.2.0",
     "torchvision==0.17.0",
@@ -19,14 +19,15 @@ dependencies = [
 packages = ["."]
 
 [tool.flwr.app]
-publisher = "flowerteam"
+publisher = "flwrlabs"
 
 [tool.flwr.app.components]
 serverapp = "fastai_example.server_app:app"
 clientapp = "fastai_example.client_app:app"
 
 [tool.flwr.app.config]
-num_server_rounds = 3
+num-server-rounds = 3
+fraction-fit = 0.5
 
 [tool.flwr.federations]
 default = "local-simulation"
